+++ order-service/src/main/java/com/odms/order/repository/ShippingFeeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShippingFeeRepository extends JpaRepository<ShippingFee, Integer> {
+public interface ShippingFeeRepository extends JpaRepository<ShippingFee, Long> {
 }
