+++ order-service/src/main/java/com/odms/order/service/impl/StatisticsServiceImpl.java
@@ -3,8 +3,8 @@ package com.odms.order.service.impl;
 import com.odms.order.dto.response.StatisticsDeliveryResponse;
 import com.odms.order.repository.OrderRepository;
 import com.odms.order.service.IStatisticsService;
-import com.odms.order.utils.WebUtils;
 import lombok.RequiredArgsConstructor;
+import nmquan.commonlib.utils.WebUtils;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -17,7 +17,7 @@ public class StatisticsServiceImpl implements IStatisticsService {
 
     @Override
     public StatisticsDeliveryResponse getDeliveryOrderStatistics(LocalDate startDate, LocalDate endDate) {
-        Integer deliveryStaffId = WebUtils.getCurrentUserId();
+        Long deliveryStaffId = WebUtils.getCurrentUserId();
         LocalDateTime startDateTime = startDate!=null ? startDate.atStartOfDay() : null;
         LocalDateTime endDateTime = endDate!=null ? endDate.atTime(23, 59, 59) : null;
         return orderRepository.getStatisticsForDeliveryStaff(deliveryStaffId, startDateTime, endDateTime)
@@ -35,7 +35,7 @@ public class StatisticsServiceImpl implements IStatisticsService {
                     response.setEarnings();
                     return response;
                 })
-                .orElse(new StatisticsDeliveryResponse(0,0,0,0,0.,
-                        0.,0,0.,0.,0.));
+                .orElse(new StatisticsDeliveryResponse(0L,0L,0L,0L,0.,
+                        0.,0L,0.,0.,0.));
     }
 }
