+++ socket-service/src/main/java/com/odms/socket/repository/SocketSessionRepository.java
@@ -5,12 +5,11 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Map;
 
 @Repository
 public interface SocketSessionRepository extends MongoRepository<SocketSession, String> {
 
     void deleteBySocketSessionId(String socketSessionId);
 
-    List<SocketSession> findAllByUserId(Integer userId);
+    List<SocketSession> findAllByUserId(Long userId);
 }
